# copa/src/main/java/com/bolao/copa/bolao/api/PalpiteResponse.java
package com.bolao.copa.bolao.api;

import java.time.Instant;

public record PalpiteResponse(
        Long id,
        JogoResponse jogo,
        Integer golsCasaPalpite,
        Integer golsForaPalpite,
        Instant createdAt,
        Instant updatedAt
) {
}